general project updates
- update dependencies
- update to C# 10.0
- fix null value parsing
- remove Parser.cs with Tomlyn's native ToModel<T> parser
- move Bindle to src/Bindle
- move Bindle.Tests to tests/Bindle.IntegrationTests
- remove Bindle/docs/readme.md; use README.md instead
