+++ .github/workflows/build-publish-release.yml
@@ -2,7 +2,7 @@ name: dotnet
 
 on:
   push:
-    tags: 
+    tags:
     - '[0-9]+.[0-9]+.[0-9]+'
     - '[0-9]+.[0-9]+.[0-9]+-preview'
     branches: [ main ]
@@ -128,7 +128,7 @@ jobs:
             echo Pushing $nupkg
             dotnet nuget push $nupkg --api-key ${{ secrets.GHPACKAGES_PAT }} --source https://nuget.pkg.github.com/${{ github.repository_owner }}/index.json --skip-duplicate
           done
-      shell: bash    
+      shell: bash
     - name: Publish Nuget Packages
       run: |
           for nupkg in $(find . -name *.nupkg)
