+++ .github/workflows/configuration.json
@@ -10,7 +10,7 @@
     {
       "title": "## 🐛 Fixes",
       "labels": [
-        "fix", 
+        "fix",
         "bug"
       ]
     },
