--- Bindle.Tests/TestPriorityAttribute.cs
@@ -1,15 +1,0 @@
-using System;
-
-namespace Deislabs.Bindle.Tests
-{
-    
-    [AttributeUsage(AttributeTargets.Method)]
-    public class TestPriorityAttribute : Attribute
-    {
-        public int Priority { get; private set; }
-        public TestPriorityAttribute(int order)
-        {
-            Priority = order;
-        }
-    }
-}
