--- Bindle/GetInvoiceOptions.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace Deislabs.Bindle
-{
-    [Flags]
-    public enum GetInvoiceOptions
-    {
-        None = 0,
-        IncludeYanked = 1,
-    }
-}
