--- Bindle/docs/readme.md
@@ -1,5 +1,0 @@
-# Bindle-dotnet
-
-This package provides a [Bindle](https://github.com/deislabs/bindle) Client for dotnet.
-
-***WARNING:*** Bindle is experimental. It is not considered production-grade by its developers, neither is it "supported" software.
