+++ README.md
@@ -1,2 +1,5 @@
 # bindle-dotnet
+
 A Bindle client for the .NET runtime
+
+***WARNING:*** Bindle is experimental. It is not considered production-grade by its developers, neither is it "supported" software.
