+++ src/Bindle/Bindle.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-    <LangVersion>9.0</LangVersion>
+    <LangVersion>10.0</LangVersion>
     <Nullable>enable</Nullable>
     <Authors>Ivan Towlson</Authors>
     <Company>Microsoft</Company>
@@ -10,16 +10,15 @@
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageId>Deislabs.Bindle</PackageId>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/deislabs/wagi-dotnet</PackageProjectUrl>
-    <PackageReadmeFile>readme.md</PackageReadmeFile>
-    <PackageReleaseNotes>https://github.com/deislabs/wagi-dotnet/releases</PackageReleaseNotes>
+    <PackageProjectUrl>https://github.com/deislabs/bindle-dotnet</PackageProjectUrl>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+    <PackageReleaseNotes>https://github.com/deislabs/bindle-dotnet/releases</PackageReleaseNotes>
     <PackageTags>Bindle;</PackageTags>
     <PublishRepositoryUrl>true</PublishRepositoryUrl>
     <RepositoryType>git</RepositoryType>
     <RootNamespace>Deislabs.Bindle</RootNamespace>
     <PublishRepositoryUrl>true</PublishRepositoryUrl>
     <EmbedUntrackedSources>true</EmbedUntrackedSources>
-
     <IncludeSymbols>true</IncludeSymbols>
     <SymbolPackageFormat>snupkg</SymbolPackageFormat>
     <MinVerDefaultPreReleasePhase>preview</MinVerDefaultPreReleasePhase>
@@ -27,18 +26,13 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Tomlyn" Version="0.1.2" />
-    <PackageReference Include="MinVer" Version="2.5.0">
+    <PackageReference Include="Tomlyn" Version="0.14.2" />
+    <PackageReference Include="MinVer" Version="3.1.0">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
-    <None Include="docs\readme.md" Pack="true" PackagePath="" />
-    <!-- See https://github.com/NuGet/Home/issues/10791 -->
-    <!-- The following reference can be removed once this issue is resolved. -->
-    <PackageReference Include="NuGet.Build.Tasks.Pack" Version="5.9.1"><IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      <PrivateAssets>all</PrivateAssets>
-    </PackageReference>
+    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1" PrivateAssets="all" />
+    <None Include="..\..\README.md" pack="true" PackagePath="." />
   </ItemGroup>
 
 </Project>
