+++ src/Bindle/GetInvoiceOptions.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Deislabs.Bindle;
+
+[Flags]
+public enum GetInvoiceOptions
+{
+    None = 0,
+    IncludeYanked = 1,
+}
