+++ tests/Bindle.IntegrationTests/data/invoices/33d7f76b3077de5b72f50d3bf1997d6571770665732ac83b3bede75cf3c5fb67/invoice.toml
