+++ tests/Bindle.IntegrationTests/data/invoices/6991cd7dd39d9e4f476a0d50da345823e7f0b76dc02326f55997158104a6efe8/invoice.toml
