+++ tests/Bindle.IntegrationTests/data/invoices/85e5c3d3a46594036dcfcf3f0097cf82bcaaa44639c93f508bb41df5f8530cad/invoice.toml
