+++ tests/Bindle.IntegrationTests/data/invoices/9acc3253d5e5be2152193fc4a4d5ee3e670e9775618942c7a66f9af5f7af1820/invoice.toml
