+++ tests/Bindle.IntegrationTests/data/invoices/a040d0a42cfe306d5ff63acfb3caddd002f70926d169bfa1d3868be4dbe1332e/invoice.toml
