+++ tests/Bindle.IntegrationTests/data/invoices/e73cf2d4f79aac38c4469f05ad6fec7c2e720357d9a7ab5dd30e95d11a658008/invoice.toml
